+++ docutils/test/test_transforms/test_transitions.py
@@ -133,33 +133,33 @@ A system message warns about invalid placement of transitions.
     <transition>
     <system_message level="2" line="1" source="test data" type="WARNING">
         <paragraph>
-            Document or section may not begin with a transition.
+            Transition at the start of the document.
     <paragraph>
         A system message warns about invalid placement of transitions.
 """],
 ["""\
-The DTD specifies ...
+rST and Doctree specifications say ...
 
 --------
 
 --------
 
-... that two transitions may not be adjacent:
+... that two transitions should not be adjacent.
 """,
 """\
 <document source="test data">
     <paragraph>
-        The DTD specifies ...
+        rST and Doctree specifications say ...
     <transition>
     <transition>
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
+            At least one body element should separate transitions.
     <paragraph>
-        ... that two transitions may not be adjacent:
+        ... that two transitions should not be adjacent.
 """],
 ["""\
-The DTD also specifies that a section or document
+The specs also say that a section or document
 may not end with a transition.
 
 --------
@@ -167,12 +167,39 @@ may not end with a transition.
 """\
 <document source="test data">
     <paragraph>
-        The DTD also specifies that a section or document
+        The specs also say that a section or document
         may not end with a transition.
     <transition>
     <system_message level="2" line="4" source="test data" type="WARNING">
         <paragraph>
-            Document may not end with a transition.
+            Transition at the end of the document.
+"""],
+["""\
+Moving and invisible elements don't count.
+
+----------
+
+.. meta:: :keywords: transition test
+.. footer:: will move away
+.. _anchor:
+.. |substitution reference| replace:: is invisible
+""",
+"""\
+<document source="test data">
+    <meta content="transition test" name="keywords">
+    <decoration>
+        <footer>
+            <paragraph>
+                will move away
+    <paragraph>
+        Moving and invisible elements don't count.
+    <transition>
+    <system_message level="2" line="3" source="test data" type="WARNING">
+        <paragraph>
+            Transition at the end of the document.
+    <target ids="anchor" names="anchor">
+    <substitution_definition names="substitution\\ reference">
+        is invisible
 """],
 ["""\
 Sections with transitions at beginning and end.
@@ -182,7 +209,7 @@ Section 1
 
 ----------
 
-Some text after transition.
+.. Comment after transition.
 
 Section 2
 =========
@@ -201,9 +228,9 @@ Some text before the transition.
         <transition>
         <system_message level="2" line="6" source="test data" type="WARNING">
             <paragraph>
-                Document or section may not begin with a transition.
-        <paragraph>
-            Some text after transition.
+                Transition at the start of the section.
+        <comment xml:space="preserve">
+            Comment after transition.
     <section ids="section-2" names="section\\ 2">
         <title>
             Section 2
@@ -212,7 +239,7 @@ Some text before the transition.
         <transition>
         <system_message level="2" line="15" source="test data" type="WARNING">
             <paragraph>
-                Document may not end with a transition.
+                Transition at the end of the document.
 """],
 ["""\
 A paragraph and two transitions.
@@ -229,10 +256,10 @@ A paragraph and two transitions.
     <transition>
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
+            At least one body element should separate transitions.
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            Document may not end with a transition.
+            Transition at the end of the document.
 """],
 ["""\
 A paragraph, two transitions, and a blank line.
@@ -250,10 +277,10 @@ A paragraph, two transitions, and a blank line.
     <transition>
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
+            At least one body element should separate transitions.
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            Document may not end with a transition.
+            Transition at the end of the document.
 """],
 ["""\
 ----------
@@ -265,46 +292,39 @@ Document beginning with a transition.
     <transition>
     <system_message level="2" line="1" source="test data" type="WARNING">
         <paragraph>
-            Document or section may not begin with a transition.
+            Transition at the start of the document.
     <paragraph>
         Document beginning with a transition.
 """],
 ["""\
+.. class:: classy
 .. meta:: :keywords: transition test
+.. footer:: will move away
+.. _anchor:
+.. |substitution reference| replace:: is invisible
 
 ----------
 
-Document beginning with a transition (meta elements don't count).
+Document beginning with a transition (title, moving elements,
+and invisible elements don't count).
 """,
 """\
 <document source="test data">
     <meta content="transition test" name="keywords">
-    <transition>
-    <system_message level="2" line="3" source="test data" type="WARNING">
-        <paragraph>
-            Document or section may not begin with a transition.
-    <paragraph>
-        Document beginning with a transition (meta elements don't count).
-"""],
-["""\
-.. header:: a header
-
-----------
-
-Document beginning with a transition (decoration elements don't count).
-""",
-"""\
-<document source="test data">
     <decoration>
-        <header>
+        <footer>
             <paragraph>
-                a header
-    <transition>
-    <system_message level="2" line="3" source="test data" type="WARNING">
+                will move away
+    <target ids="anchor" names="anchor">
+    <substitution_definition names="substitution\\ reference">
+        is invisible
+    <transition classes="classy">
+    <system_message level="2" line="7" source="test data" type="WARNING">
         <paragraph>
-            Document or section may not begin with a transition.
+            Transition at the start of the document.
     <paragraph>
-        Document beginning with a transition (decoration elements don't count).
+        Document beginning with a transition (title, moving elements,
+        and invisible elements don't count).
 """],
 ["""\
 Section 1
@@ -316,14 +336,8 @@ Section 1
 
 ----------
 
-Implementation Detail
-=====================
-
-If the element containing the transition is invalid after replacing the
-transition with a body element, the system_message is appended at the end
-of the document (by the "universal.Messages" transform).
-This check can lead to overcautious behaviour if there are other
-validity violations (here: several misplaced transitions).
+Section 2
+=========
 """,
 """\
 <document source="test data">
@@ -331,26 +345,17 @@ validity violations (here: several misplaced transitions).
         <title>
             Section 1
         <transition>
+        <system_message level="2" line="4" source="test data" type="WARNING">
+            <paragraph>
+                Transition at the start of the section.
         <transition>
     <transition>
-    <section ids="implementation-detail" names="implementation\\ detail">
-        <title>
-            Implementation Detail
-        <paragraph>
-            If the element containing the transition is invalid after replacing the
-            transition with a body element, the system_message is appended at the end
-            of the document (by the "universal.Messages" transform).
-            This check can lead to overcautious behaviour if there are other
-            validity violations (here: several misplaced transitions).
-    <system_message level="2" line="4" source="test data" type="WARNING">
-        <paragraph>
-            Document or section may not begin with a transition.
-    <system_message level="2" line="6" source="test data" type="WARNING">
-        <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
     <system_message level="2" line="8" source="test data" type="WARNING">
         <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
+            At least one body element should separate transitions.
+    <section ids="section-2" names="section\\ 2">
+        <title>
+            Section 2
 """],
 ["""\
 ----------
@@ -364,35 +369,17 @@ validity violations (here: several misplaced transitions).
 """\
 <document source="test data">
     <transition>
-    <transition>
-    <transition>
-    <system_message level="2" line="5" source="test data" type="WARNING">
-        <paragraph>
-            Document may not end with a transition.
     <system_message level="2" line="1" source="test data" type="WARNING">
         <paragraph>
-            Document or section may not begin with a transition.
-    <system_message level="2" line="3" source="test data" type="WARNING">
-        <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
+            Transition at the start of the document.
+    <transition>
+    <transition>
     <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            At least one body element must separate transitions; adjacent transitions are not allowed.
-"""],
-["""\
-A paragraph.
-
-----------
-
-""",
-"""\
-<document source="test data">
-    <paragraph>
-        A paragraph.
-    <transition>
-    <system_message level="2" line="3" source="test data" type="WARNING">
+            At least one body element should separate transitions.
+    <system_message level="2" line="5" source="test data" type="WARNING">
         <paragraph>
-            Document may not end with a transition.
+            Transition at the end of the document.
 """],
 ])
 
@@ -425,7 +412,7 @@ totest_x['transitions extra'] = ((Transitions,), [
         <transition>
         <system_message level="2" line="7" source="test data" type="WARNING">
             <paragraph>
-                Transition must be child of <document> or <section>.
+                Transition only valid as child of <document> or <section>.
         <paragraph>
             Some text.
     <paragraph>
@@ -434,7 +421,7 @@ totest_x['transitions extra'] = ((Transitions,), [
          in a paragraph.
     <system_message level="2" line="10" source="test data" type="WARNING">
         <paragraph>
-            Transition must be child of <document> or <section>.
+            Transition only valid as child of <document> or <section>.
 """],
 ])